add more conf files

+++ vagrant_shell/deb.sh
@@ -6,6 +6,7 @@ install_java=true
 
 install_hadoop=true
 install_hive=true
+install_hive2=true
 install_confluent=true
 install_flink=true
 install_mongo=true
@@ -20,11 +21,14 @@ install_phoenix=true
 #software repository links
 dl_link_java=http://download.oracle.com/otn-pub/java/jdk/8u161-b12/2f38c3b165be4555a1fa6e98c45e0808/jdk-8u161-linux-x64.tar.gz
 
-file_name_hadoop=hadoop-2.6.0.tar.gz
-dl_link_hadoop=https://archive.apache.org/dist/hadoop/common/hadoop-2.6.0/hadoop-2.6.0.tar.gz
+file_name_hadoop=hadoop-2.7.3.tar.gz
+dl_link_hadoop=https://archive.apache.org/dist/hadoop/common/hadoop-2.7.3/hadoop-2.7.3.tar.gz
 
 file_name_hive=hive-1.2.1.tar.gz
 dl_link_hive=https://archive.apache.org/dist/hive/hive-1.2.1/apache-hive-1.2.1-bin.tar.gz
+
+file_name_hive2=hive-2.3.2.tar.gz
+dl_link_hive2=https://archive.apache.org/dist/hive/hive-2.3.2/apache-hive-2.3.2-bin.tar.gz
 
 file_name_confluent=confluent-3.3.0.tar.gz
 dl_link_confluent=http://packages.confluent.io/archive/3.3/confluent-oss-3.3.0-2.11.tar.gz
@@ -41,14 +45,14 @@ dl_link_zeppelin=https://archive.apache.org/dist/zeppelin/zeppelin-0.7.3/zeppeli
 file_name_grafana=grafana-4.6.2.tar.gz
 dl_link_grafana=https://s3-us-west-2.amazonaws.com/grafana-releases/release/grafana-4.6.2.linux-x64.tar.gz
 
-file_name_spark=spark-2.3.0.tgz
-dl_link_spark=https://archive.apache.org/dist/spark/spark-2.3.0/spark-2.3.0-bin-hadoop2.6.tgz
+file_name_spark=spark-2.2.0.tgz
+dl_link_spark=https://archive.apache.org/dist/spark/spark-2.2.0/spark-2.2.0-bin-hadoop2.7.tgz
 
-file_name_hbase=hbase-1.3.0.tar.gz
-dl_link_hbase=https://archive.apache.org/dist/hbase/1.3.0/hbase-1.3.0-bin.tar.gz
+file_name_hbase=hbase-1.1.2.tar.gz
+dl_link_hbase=https://archive.apache.org/dist/hbase/1.1.2/hbase-1.1.2-bin.tar.gz
 
-file_name_phoenix=apache-phoenix-4.13.1-HBase-1.3-bin.tar.gz
-dl_link_phoenix=http://archive.apache.org/dist/phoenix/apache-phoenix-4.13.1-HBase-1.3/bin/apache-phoenix-4.13.1-HBase-1.3-bin.tar.gz
+file_name_phoenix=phoenix-4.13.1.tar.gz
+dl_link_phoenix=http://archive.apache.org/dist/phoenix/apache-phoenix-4.13.1-HBase-1.1/bin/apache-phoenix-4.13.1-HBase-1.1-bin.tar.gz
 
 file_name_livy=livy-0.4.0.tar.gz
 dl_link_livy=https://github.com/datafibers-community/df_demo/releases/download/livy/livy-0.4.0-incubating-bin.tar.gz
@@ -120,6 +124,9 @@ soft_install $install_hadoop hadoop $dl_link_hadoop $file_name_hadoop
 
 # Install and configure Hive
 soft_install $install_hive hive $dl_link_hive $file_name_hive
+
+# Install and configure Hive v2
+soft_install $install_hive2 hive2 $dl_link_hive2 $file_name_hive2
 
 # Install CP
 soft_install $install_confluent confluent $dl_link_confluent $file_name_confluent
@@ -232,30 +239,40 @@ if [ "$install_zeppelin" = true ]; then
   cp /opt/hadoop/share/hadoop/common/hadoop-common-*.jar /opt/zeppelin/interpreter/jdbc/
 fi
 
-# Install MySQL Metastore for Hive - do this after creating profiles in order to use hive schematool
-sudo apt-get -y update
-sudo debconf-set-selections <<< 'mysql-server mysql-server/root_password password mypassword'
-sudo debconf-set-selections <<< 'mysql-server mysql-server/root_password_again password mypassword'
-sudo apt-get -y install mysql-server
-sudo apt-get -y install libmysql-java
+if [ "$install_hive" = true ] || [ "$install_hive2" = true ]; then
+    # Install MySQL Metastore for Hive - do this after creating profiles in order to use hive schematool
+    sudo apt-get -y update
+    sudo debconf-set-selections <<< 'mysql-server mysql-server/root_password password mypassword'
+    sudo debconf-set-selections <<< 'mysql-server mysql-server/root_password_again password mypassword'
+    sudo apt-get -y install mysql-server
+    sudo apt-get -y install libmysql-java
 
-sudo ln -sfn /usr/share/java/mysql-connector-java.jar /opt/hive/lib/mysql-connector-java.jar
+    # Configure Hive Metastore
+    mysql -u root --password="mypassword" -f \
+    -e "DROP DATABASE IF EXISTS metastore;"
 
-# Configure Hive Metastore
-mysql -u root --password="mypassword" -f \
--e "DROP DATABASE IF EXISTS metastore;"
+    mysql -u root --password="mypassword" -f \
+    -e "CREATE DATABASE IF NOT EXISTS metastore;"
 
-mysql -u root --password="mypassword" -f \
--e "CREATE DATABASE IF NOT EXISTS metastore;"
+    mysql -u root --password="mypassword" \
+    -e "GRANT ALL PRIVILEGES ON metastore.* TO 'hive'@'localhost' IDENTIFIED BY 'mypassword'; FLUSH PRIVILEGES;"
 
-mysql -u root --password="mypassword" \
--e "GRANT ALL PRIVILEGES ON metastore.* TO 'hive'@'localhost' IDENTIFIED BY 'mypassword'; FLUSH PRIVILEGES;"
-
-schematool -dbType mysql -initSchema
+    if [ "$install_hive2" = true ]; then
+        sudo ln -sfn /usr/share/java/mysql-connector-java.jar /opt/hive2/lib/mysql-connector-java.jar
+        cp /mnt/etc/hive2/hive-site.xml /opt/hive2/conf/
+        /opt/hive2/bin/schematool -dbType mysql -initSchema
+        echo "Init. schema using hive version 2"
+    else
+        sudo ln -sfn /usr/share/java/mysql-connector-java.jar /opt/hive/lib/mysql-connector-java.jar
+        cp /mnt/etc/hive/hive-site.xml /opt/hive/conf/
+        /opt/hive/bin/schematool -dbType mysql -initSchema
+        echo "Init. schema using hive version 1"
+    fi
+fi
 
 echo "***************************************************************************************"
-echo "* Lab Virtual Machine Setup Completed.                                                 *"
-echo "* SSH address:127.0.0.1:2222.                                                          *"
+echo "* Lab Virtual Machine Setup is Completed.                                              *"
+echo "* SSH address:localhost:2222.                                                          *"
 echo "* SSH username/password:vagrant/vagrant                                                *"
 echo "* Command: ssh vagrant@localhost -p 2222                                               *"
 echo "***************************************************************************************"
